broadcom-diag: add Belkin F7Dxxxx
While marked and marketed as different boards, the F7Dxxxx series is
(so far) largely the same in what matters for broadcom-diag: all the
buttons and leds are assigned to the same gpios.

The only difference that may need to be noted: some models lack a usb1
led (as they lack usb1). Given that it is relatively easy to populate
that led, I see no point in disabling access to it.

Signed-off-by: Cody P Schafer <devel@codyps.com>