broadcom-diag: tweak Belkin F7Dxxxx, share 1 led/button layout
While marked and marketed as different boards, the F7Dxxxx series is
(so far) largely the same in what matters for broadcom-diag: all the
buttons and leds are assigned to the same gpios.

The only difference that may need to be noted: some models lack a usb1
led (as they lack usb1). Given that it is relatively easy to populate
that led, I see no point in disabling access to it. Additionally, this
does mean that /proc/diag/model will only show the non-specific model.
As /proc/cpuinfo's "machine:" entry still shows the detailed board
type, we don't loose anything overall.

Signed-off-by: Cody P Schafer <devel@codyps.com>