brcm47xx/mtd-partition: recognize Belkin F7Dxxxx TRX headers
F7D3301/F7D7301, F7D3302/F7D7302, and F7D4301 replace the "HDR0" with a
custom byte sequence (which appears to be a date). Also, they all
appear to support a single "QA" byte sequence.

The QA byte sequence doesn't appear to work via the webinterface, so we
need to support all of them.

Signed-off-by: Cody P Schafer <devel@codyps.com>